requirements.txt
aistudio-sdk==0.3.8
alembic==1.17.2
annotated-types==0.7.0
anyio==4.12.1
astor==0.8.1
attrdict==2.0.1
babel==2.17.0
bce-python-sdk==0.9.59
beautifulsoup4==4.12.2
blinker==1.9.0
cachetools==6.2.4
certifi==2026.1.4
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
colorlog==6.10.1
contourpy==1.3.3
cssselect==1.3.0
cssutils==2.11.1
cycler==0.12.1
Cython==3.2.4
decorator==5.2.1
et_xmlfile==2.0.0
filelock==3.20.3
fire==0.7.1
Flask==3.0.0
flask-babel==4.0.0
Flask-Cors==4.0.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
fonttools==4.61.1
fsspec==2026.1.0
future==1.0.0
greenlet==3.3.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.1
idna==3.11
ImageIO==2.37.2
imagesize==1.4.1
imgaug==0.4.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
lazy_loader==0.4
lmdb==1.7.5
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.8
modelscope==1.33.0
more-itertools==10.8.0
networkx==3.6.1
nltk==3.8.1
numpy==1.26.4
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
opt-einsum==3.3.0
packaging==25.0
paddleocr==2.7.3
paddlepaddle==2.6.2
paddlex==3.3.13
pandas==2.3.3
pdf2docx==0.5.8
Pillow==10.1.0
playwright==1.57.0
premailer==3.10.0
prettytable==3.17.0
protobuf==3.20.2
psutil==7.2.1
py-cpuinfo==9.0.0
pyclipper==1.4.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.0
PyJWT==2.8.0
PyMuPDF==1.26.7
pyodbc==5.2.0
pyparsing==3.3.1
pypdfium2==5.3.0
pytesseract==0.3.10
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.14.3
rarfile==4.2
regex==2025.11.3
requests==2.31.0
ruamel.yaml==0.19.1
scikit-image==0.26.0
scipy==1.17.0
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
soupsieve==2.8.1
SQLAlchemy==2.0.45
termcolor==3.3.0
tifffile==2025.12.20
tqdm==4.67.1
typer-slim==0.21.1
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
ujson==5.11.0
urllib3==2.6.3
visualdl==2.5.3
wcwidth==0.2.14
Werkzeug==3.1.4
wheel==0.45.1